Do not url quote thumb url
This was resulting in the storage's exists method returning an incorrect
false result because the quoted file key was then encoded again.
Because it could never determine the existance of files with special
characters, the template tag would recreate thumbnails for those images
on every invocation. This was the cause of a huge latency spike.